+++ .gitignore
@@ -27,3 +27,7 @@ node_modules
 # Generated sources
 /generated
 src/generated/
+/*.d.ts
+/*.d.ts.map
+/*.js
+/*.js.map
