--- src/Checkbox.ts
@@ -1,1 +1,0 @@
-export * from "./generated/Checkbox.js";
