+++ src/Checkbox.tsx
@@ -1,0 +1,19 @@
+import { type ForwardedRef, forwardRef, type ReactElement } from 'react';
+import {
+  Checkbox as _Checkbox,
+  CheckboxModule,
+  type CheckboxProps as _CheckboxProps,
+} from './generated/Checkbox.js';
+
+export type CheckboxProps = _CheckboxProps &
+  Readonly<{
+    children?: _CheckboxProps['label'];
+  }>;
+
+function Checkbox(props: CheckboxProps, ref: ForwardedRef<CheckboxModule.Checkbox>): ReactElement | null {
+  return <_Checkbox {...props} ref={ref} label={props.label ?? props.children} />;
+}
+
+const ForwardedCheckbox = forwardRef(Checkbox);
+
+export { ForwardedCheckbox as Checkbox, CheckboxModule };
