+++ test/kitchen-sink/data.ts
@@ -1,0 +1,14 @@
+enum CrudRole {
+  ADMIN = "admin",
+  USER = "user",
+}
+
+export type CrudDataItem = {
+  name: string;
+  role: CrudRole;
+  active: boolean;
+};
+export const crudData: CrudDataItem[] = [
+  { name: "Jane Doe", role: CrudRole.ADMIN, active: true },
+  { name: "Mary Joe", role: CrudRole.USER, active: false }
+];
